Fix eslint errors in tests
